requirements.txt
conekta
Adyen==1.2.0
algoliasearch==1.15.3
amqp==2.2.1
anyjson==0.3.3
argon2-cffi==16.3.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
Django==1.11.7
django-cors-headers==2.1.0
django-environ==0.4.3
django-filter==1.0.4
djangorestframework==3.7.0
djangorestframework-jwt==1.11.0
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
hiredis==0.2.0
idna==2.5
kombu==4.1.0
Markdown==2.6.8
mysqlclient==1.3.10
newrelic==2.100.0.84
pycparser==2.17
Pygments==2.2.0
PyJWT==1.5.2
python-dateutil==2.6.0
pytz==2017.3
PyYAML==3.12
redis==2.10.5
requests==2.18.4
six==1.10.0
urllib3==1.21.1
vine==1.1.4
